requirements.txt
absl-py==0.9.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.4
blinker==1.4
boto==2.49.0
boto3==1.14.6
botocore==1.17.6
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2.4
docutils==0.15.2
editdistance==0.5.3
entrypoints==0.3
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.1
future==0.18.2
gast==0.3.3
gensim==3.8.0
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==19.9.0
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
imageio==2.8.0
importlib-metadata==1.6.0
importlib-resources==1.4.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.11
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==2.1.1
mistune==0.8.4
mock==4.0.2
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
npm==0.1.1
numpy==1.18.5
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
optional-django==0.1.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
pluggy==0.6.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.4.2
ptyprocess==0.6.0
py==1.5.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
pyrsistent==0.16.0
pytest==3.3.2
python-apt==1.6.5+ubuntu0.3
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-debian==0.1.32
python-xapp==1.0.0
python-xlib==0.20
pytz==2019.3
pyvi==0.0.9.7
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.0
q==2.6
qtconsole==4.7.3
QtPy==1.9.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.19.1
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.14.0
sklearn-crfsuite==0.3.6
smart-open==2.0.0
sqlparse==0.3.1
ssh-import-id==5.7
stevedore==1.32.0
system-service==0.3
systemd-python==234
tabulate==0.8.7
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.6.3
torch==1.5.0+cpu
torchvision==0.6.0+cpu
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.22
virtualenv==20.0.18
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
